+++ packages/components/organisms/f-checkout/src/components/_tests/ErrorDialog.test.js
@@ -12,23 +12,30 @@ const localVue = createLocalVue();
 localVue.use(VueI18n);
 localVue.use(Vuex);
 
-let windowLocationSpy;
+const propsData = {};
+const restaurant = {
+    seoName: 'checkout-kofte-farringdon',
+    id: '22222'
+};
+const defaultMessage = {
+    shouldRedirectToMenu: true,
+    shouldShowInDialog: true,
+    code: ERROR_CODE_FULFILMENT_TIME_UNAVAILABLE
+};
+const duplicateOrderMessage = {
+    code: 'DuplicateOrder'
+};
 
-beforeEach(() => {
-    windowLocationSpy = jest.spyOn(window.location, 'assign').mockImplementation();
+afterEach(() => {
+    jest.clearAllMocks();
 });
 
 describe('ErrorDialog', () => {
-    const propsData = {};
-    const restaurant = {
-        seoName: 'checkout-kofte-farringdon',
-        id: '22222'
-    };
-    const message = {
-        shouldRedirectToMenu: true,
-        shouldShowInDialog: true,
-        code: ERROR_CODE_FULFILMENT_TIME_UNAVAILABLE
-    };
+    let windowLocationSpy;
+
+    beforeEach(() => {
+        windowLocationSpy = jest.spyOn(window.location, 'assign').mockImplementation();
+    });
 
     it('should be defined', () => {
         // Arrange && Act
@@ -50,7 +57,7 @@ describe('ErrorDialog', () => {
                 const wrapper = shallowMount(ErrorDialog, {
                     store: createStore({
                         ...defaultCheckoutState,
-                        message
+                        message: defaultMessage
                     }),
                     i18n,
                     localVue,
@@ -58,7 +65,7 @@ describe('ErrorDialog', () => {
                 });
 
                 // Assert
-                expect(wrapper.vm.errorCode).toEqual(message.code);
+                expect(wrapper.vm.errorCode).toEqual(defaultMessage.code);
             });
 
             it('should return `null` if  no `message` is provided', () => {
@@ -102,10 +109,6 @@ describe('ErrorDialog', () => {
                 updateMessageSpy = jest.spyOn(ErrorDialog.methods, 'updateMessage');
             });
 
-            afterEach(() => {
-                jest.clearAllMocks();
-            });
-
             it('should call `updateMessage`', () => {
                 // Arrange
                 const wrapper = shallowMount(ErrorDialog, {
@@ -128,7 +131,7 @@ describe('ErrorDialog', () => {
                     store: createStore({
                         ...defaultCheckoutState,
                         restaurant,
-                        message
+                        message: defaultMessage
                     }),
                     i18n,
                     localVue,
@@ -144,13 +147,13 @@ describe('ErrorDialog', () => {
 
             it('should not redirect to the restaurant menu if `shouldRedirectToMenu` is false', () => {
                 // Arrange
-                message.shouldRedirectToMenu = false;
+                defaultMessage.shouldRedirectToMenu = false;
 
                 const wrapper = shallowMount(ErrorDialog, {
                     store: createStore({
                         ...defaultCheckoutState,
                         restaurant,
-                        message
+                        message: defaultMessage
                     }),
                     i18n,
                     localVue,
@@ -164,5 +167,164 @@ describe('ErrorDialog', () => {
                 expect(windowLocationSpy).not.toHaveBeenCalled();
             });
         });
+
+        describe('showOrderHistory ::', () => {
+            let updateMessageSpy;
+
+            beforeEach(() => {
+                updateMessageSpy = jest.spyOn(ErrorDialog.methods, 'updateMessage');
+            });
+
+            it('should call `updateMessage`', () => {
+                // Arrange
+                const wrapper = shallowMount(ErrorDialog, {
+                    store: createStore(),
+                    i18n,
+                    localVue,
+                    propsData
+                });
+
+                // Act
+                wrapper.vm.closeErrorDialog();
+
+                // Assert
+                expect(updateMessageSpy).toHaveBeenCalled();
+            });
+
+            it('should redirect to the order history', () => {
+                // Arrange
+                const wrapper = shallowMount(ErrorDialog, {
+                    store: createStore({
+                        ...defaultCheckoutState,
+                        restaurant,
+                        message: duplicateOrderMessage
+                    }),
+                    i18n,
+                    localVue,
+                    propsData
+                });
+
+                // Act
+                wrapper.vm.showOrderHistory();
+
+                // Assert
+                expect(windowLocationSpy).toHaveBeenCalledWith('order-history');
+            });
+        });
+
+        describe('isDuplicateOrderError ::', () => {
+            it('should return `True` when `ErrorCode` refers to a duplicate order', () => {
+                // Arrange
+                const wrapper = shallowMount(ErrorDialog, {
+                    store: createStore({
+                        ...defaultCheckoutState,
+                        message: duplicateOrderMessage
+                    }),
+                    i18n,
+                    localVue,
+                    propsData
+                });
+
+                // Act
+                const actual = wrapper.vm.isDuplicateOrderError();
+
+                // Assert
+                expect(actual).toBeTruthy();
+            });
+
+            it('should return `False` when `ErrorCode` does not refer to a duplicate order', () => {
+                // Arrange
+                const wrapper = shallowMount(ErrorDialog, {
+                    store: createStore({
+                        ...defaultCheckoutState,
+                        message: defaultMessage
+                    }),
+                    i18n,
+                    localVue,
+                    propsData
+                });
+
+                // Act
+                const actual = wrapper.vm.isDuplicateOrderError();
+
+                // Assert
+                expect(actual).toBeFalsy();
+            });
+        });
+
+        describe('dataLayerPushDupOrderWarnTrackingEvent ::', () => {
+            it('should push to the dataLayer', () => {
+                // Arrange
+                const expected = {
+                    event: 'trackEvent',
+                    eventCategory: 'engagement',
+                    eventAction: 'dialog_duplicate_order_warning',
+                    eventLabel: 'view_dialog'
+                };
+
+                Object.defineProperty(global, 'window', {
+                    value: {
+                        dataLayer: []
+                    }
+                });
+                window.dataLayer = [];
+
+                const wrapper = shallowMount(ErrorDialog, {
+                    store: createStore({
+                        ...defaultCheckoutState,
+                        message: duplicateOrderMessage
+                    }),
+                    i18n,
+                    localVue,
+                    propsData
+                });
+
+                // Act
+                wrapper.vm.dataLayerPushDupOrderWarnTrackingEvent();
+
+                // Assert
+                expect(window.dataLayer).toContainEqual(expected);
+            });
+        });
+    });
+});
+
+describe('ErrorDialog :: mounted ::', () => {
+    let dataLayerPushDupOrderWarnTrackingEventSpy;
+
+    beforeEach(() => {
+        dataLayerPushDupOrderWarnTrackingEventSpy = jest.spyOn(ErrorDialog.methods, 'dataLayerPushDupOrderWarnTrackingEvent');
+    });
+
+    it('should make a call to `dataLayerPushDupOrderWarnTrackingEvent` when `ErrorCode` refers to a duplicate order ', () => {
+        // Arrange & Act
+        shallowMount(ErrorDialog, {
+            store: createStore({
+                ...defaultCheckoutState,
+                message: duplicateOrderMessage
+            }),
+            i18n,
+            localVue,
+            propsData
+        });
+
+        // Assert
+        expect(dataLayerPushDupOrderWarnTrackingEventSpy).toHaveBeenCalled();
+    });
+
+    it('should not make a call to `dataLayerPushDupOrderWarnTrackingEvent` when `ErrorCode` does not refer to a duplicate order ', () => {
+        // Arrange & Act
+        shallowMount(ErrorDialog, {
+            store: createStore({
+                ...defaultCheckoutState,
+                message: defaultMessage
+            }),
+            i18n,
+            localVue,
+            propsData
+        });
+
+        // Assert
+        expect(dataLayerPushDupOrderWarnTrackingEventSpy).not.toHaveBeenCalled();
     });
 });