+++ packages/components/organisms/f-checkout/src/tenants/en-GB.js
@@ -116,7 +116,8 @@ const messages = {
             DuplicateOrder: {
                 title: 'This looks like a duplicate order',
                 message: 'So you’re not charged twice, your order has not been processed',
-                buttonText: 'Ok'
+                buttonText: 'Close',
+                buttonTextPrimary: 'View my orders'
             }
         }
     },