+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -4,6 +4,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.127.0
+------------------------------
+*June 1, 2021*
+
+### Change
+- Give the consumer two options; 'Close' or 'View recent orders', when showing the Duplicate Error Warning dialog box
+- Added tracking to the alternative dialog box
+
+
 v0.121.0
 ------------------------------
 *May 24, 2021*